moving useOnClickAuthenticateWithGithub to its own file
