+++ editor/src/components/github/github-repository-clone-flow.tsx
@@ -4,7 +4,7 @@ import { forceNotNull } from '../../core/shared/optional-utils'
 import { NO_OP } from '../../core/shared/utils'
 import { totallyEmptyDefaultProject } from '../../sample-projects/sample-project-utils'
 import invariant from '../../third-party/remix/invariant'
-import { useOnClickAuthenticateWithGithub } from '../../utils/github-auth'
+import { useOnClickAuthenticateWithGithub } from '../../utils/github-auth-hooks'
 import { Dialog, FormButton } from '../../uuiui'
 import { isLoggedIn, type EditorDispatch } from '../editor/action-types'
 import { setGithubState } from '../editor/actions/action-creators'
