+++ editor/src/utils/github-auth-hooks.ts
@@ -1,0 +1,11 @@
+import React from 'react'
+import { useDispatch } from '../components/editor/store/dispatch-context'
+import { GithubAuth } from './github-auth'
+
+export function useOnClickAuthenticateWithGithub() {
+  const dispatch = useDispatch()
+  const triggerAuthentication = React.useCallback(() => {
+    void GithubAuth.startGithubAuthentication(dispatch)
+  }, [dispatch])
+  return triggerAuthentication
+}
